2022년 3월 17일 Board Update/Delete 세션 아이디 처리

+++ MyfirstWeb/src/main/java/kr/co/ict/servlet/service/BoardDeleteService.java
@@ -1,7 +1,9 @@
 package kr.co.ict.servlet.service;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.BoardDAO;
 
@@ -9,11 +11,15 @@ public class BoardDeleteService implements IBoardService{
 
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) {
+		HttpSession session = request.getSession();
+		String sId = (String)session.getAttribute("userId");
+		String writer = request.getParameter("writer");
 		String strNum = request.getParameter("board_num");
 		int bNum = Integer.parseInt(strNum);
 		
-		BoardDAO dao = BoardDAO.getInstance();
-		dao.deleteBoard(bNum);		
+		if(sId.equals(writer)) {
+			BoardDAO dao = BoardDAO.getInstance();
+			dao.deleteBoard(bNum);		
+		} 
 	}
-
 }
